added random team functionality
