+++ pages/team/style.module.scss
@@ -79,3 +79,7 @@
     color: $color-brand-negative
   }
 }
+
+.addButtons button{
+  margin-right: 1em;
+}
