+++ src/common/actions/getCP.ts
@@ -1,10 +1,13 @@
-import { cpms } from "@config/statVals"
+import { cpms } from '@config/statVals'
 
-const getCP = (bs: {
-  atk: number,
-  def: number,
-  hp: number
-}, stats: number[]): number => {
+const getCP = (
+  bs: {
+    atk: number
+    def: number
+    hp: number
+  },
+  stats: number[]
+): number => {
   const cpm = cpms[(stats[0] - 1) * 2]
   const atk = stats[1]
   const def = stats[2]
